refactor: mandara.js を責務分離してファイルサイズ制限対応
Code Quality Check (max 1000 lines) 対応:
mandara.js 1380行 → 899行 (-35%)

新規モジュール:
- js/mandara-insight-controller.js (約370行):
  Insight機能のフロー制御 (showInsightIntro, startInsightAnalysis,
  startCrossAnalysis, handleInsightCreateGarage 等) を切り出し。
  context オブジェクト経由で状態・副作用を注入 (依存性逆転)
- js/mandara-tags-todos-ui.js (約170行):
  タグ/TODOの描画・追加・編集・削除・並び替えUIを切り出し。
  同じく context 経由で currentMandara 参照と保存を受け取る

mandara.js 側の変更:
- Insight関連の全関数を controller 呼び出しに置き換え
- Tags/Todos 関連の関数を薄いラッパー関数に変更
  (同じ関数名を維持することでイベントハンドラの呼び出しを変更せず済む)
- 冗長な import を整理
